used picocli to organize convert/execute/version commands
