+++ vcell-cli/pom.xml
@@ -79,6 +79,11 @@
 
     <dependencies>
         <dependency>
+            <groupId>info.picocli</groupId>
+            <artifactId>picocli</artifactId>
+            <version>4.6.3</version>
+        </dependency>
+        <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
             <scope>test</scope>
@@ -162,6 +167,8 @@
             <artifactId>commons-io</artifactId>
             <version>2.7</version>
         </dependency>
+
+
     </dependencies>
     <profiles>
         <profile>
