+++ vcell-cli/src/main/java/org/vcell/cli/vcml/ModelFormat.java
@@ -1,0 +1,5 @@
+package org.vcell.cli.vcml;
+
+public enum ModelFormat {
+    SBML, VCML;
+}
