+++ vcell-cli/src/main/java/org/vcell/cli/vcml/VcmlOmexConverter.java
@@ -34,9 +34,7 @@ import org.sbpax.schemas.BioPAX3;
 import org.sbpax.schemas.util.DefaultNameSpaces;
 import org.sbpax.schemas.util.OntUtil;
 import org.sbpax.util.SesameRioUtil;
-import org.vcell.cli.CLIDatabaseService;
-import org.vcell.cli.CLIHandler;
-import org.vcell.cli.CLIStandalone;
+import org.vcell.cli.*;
 import org.vcell.sedml.PubMet;
 import org.vcell.sedml.SEDMLExporter;
 import org.vcell.util.DataAccessException;
@@ -48,28 +46,32 @@ import java.io.FileNotFoundException;
 import java.io.FilenameFilter;
 import java.io.IOException;
 import java.net.URL;
+import java.nio.file.Files;
 import java.nio.file.Paths;
+import java.nio.file.StandardOpenOption;
 import java.sql.SQLException;
 import java.text.SimpleDateFormat;
 import java.util.*;
 
+
+
 public class VcmlOmexConverter {
 
 	// TODO: Changes to CLIHandler have made some things here superfluous; this can be trimmed(?).
-	public static void convertFiles(CLIDatabaseService cliDatabaseService, CLIHandler cliHandler) throws IOException, SQLException, DataAccessException {
-        File input = null;
-        try {
-            // TODO: handle if it's not valid PATH (NOTE: Changes to CLIHander should make this moot!)
-            input = new File(cliHandler.getInputFilePath());
-        } catch (Exception e1) {
-            e1.printStackTrace();
-        }
-
-		boolean bForceSBML = cliHandler.shouldForceSbml();
-		boolean bForceVCML = cliHandler.shouldForceVcml();
-		boolean bHasDataOnly = cliHandler.isHasDataOnly();
-		boolean bMakeLogsOnly = cliHandler.isMakeLogsOnly();
-		boolean bNonSpatialOnly = cliHandler.isNonSpacialOnly();
+	public static void convertFiles(CLIDatabaseService cliDatabaseService,
+									File input,
+									File outputDir,
+									ModelFormat modelFormat,
+									boolean bHasDataOnly,
+									boolean bMakeLogsOnly,
+									boolean bNonSpatialOnly,
+									boolean bForceLogFiles)
+			throws IOException, SQLException, DataAccessException {
+//		boolean bForceSBML = cliHandler.shouldForceSbml();
+//		boolean bForceVCML = cliHandler.shouldForceVcml();
+//		boolean bHasDataOnly = cliHandler.isHasDataOnly();
+//		boolean bMakeLogsOnly = cliHandler.isMakeLogsOnly();
+//		boolean bNonSpatialOnly = cliHandler.isNonSpacialOnly();
 
 		VCInfoContainer vcic;
 		Map<String, List<String>> publicationToModelMap = new LinkedHashMap<>();
@@ -111,7 +113,7 @@ public class VcmlOmexConverter {
 				for(String model : models) {
 					row += (", " + model);
 				}
-				CLIStandalone.writeMultiModelPublications(cliHandler.getOutputDirPath(), row);
+				writeMultiModelPublications(outputDir.getAbsolutePath(), row, bForceLogFiles);
 				System.out.println(row);
 			}
 		}
@@ -124,13 +126,12 @@ public class VcmlOmexConverter {
             if (inputFiles == null) {
                 System.err.println("No VCML files found in the directory `" + input + "`");
             }
-            String outputDir = cliHandler.getOutputDirPath();			// full directory name, like C:\TEMP\biomodel\omex\native
-            
-            CLIStandalone.writeSimErrorList(outputDir, "bForceVCML is " + bForceVCML);
-            CLIStandalone.writeSimErrorList(outputDir, "bForceSBML is " + bForceSBML);
-            CLIStandalone.writeSimErrorList(outputDir, "hasDataOnly is " + bHasDataOnly);
-            CLIStandalone.writeSimErrorList(outputDir, "makeLogsOnly is " + bMakeLogsOnly);
-            CLIStandalone.writeSimErrorList(outputDir, "nonSpatialOnly is " + bNonSpatialOnly);
+
+            writeSimErrorList(outputDir.getAbsolutePath(), "bForceVCML is " + modelFormat.equals(ModelFormat.VCML), bForceLogFiles);
+			writeSimErrorList(outputDir.getAbsolutePath(), "bForceSBML is " + modelFormat.equals(ModelFormat.SBML), bForceLogFiles);
+			writeSimErrorList(outputDir.getAbsolutePath(), "hasDataOnly is " + bHasDataOnly, bForceLogFiles);
+			writeSimErrorList(outputDir.getAbsolutePath(), "makeLogsOnly is " + bMakeLogsOnly, bForceLogFiles);
+			writeSimErrorList(outputDir.getAbsolutePath(), "nonSpatialOnly is " + bNonSpatialOnly, bForceLogFiles);
 
 //            assert inputFiles != null;
             for (String inputFile : inputFiles) {
@@ -138,8 +139,8 @@ public class VcmlOmexConverter {
                 System.out.println(" ============== start: " + inputFile);
                 try {
                     if (inputFile.endsWith(".vcml")) {
-                        boolean isCreated = vcmlToOmexConversion(file.toString(), outputDir, cliDatabaseService,
-								cliHandler, hasNonSpatialSet, hasSpatialSet);
+                        boolean isCreated = vcmlToOmexConversion(file.toString(), outputDir.getAbsolutePath(), outputDir.getAbsolutePath(), cliDatabaseService,
+								hasNonSpatialSet, hasSpatialSet, modelFormat, bHasDataOnly, bNonSpatialOnly, bMakeLogsOnly, bForceLogFiles);
                         if (isCreated) {
                         	System.out.println("Combine archive created for file(s) `" + inputFile + "`");
                         }
@@ -153,20 +154,27 @@ public class VcmlOmexConverter {
 //					e.printStackTrace(System.err);
                     
                 	System.out.println("\n\n\n=====>>>>EXPORT FAILED: " +inputFile+"\n\n\n");
-                	CLIStandalone.writeDetailedErrorList(outputDir, inputFile + ",   " + e.getMessage());
+					CLIUtils.writeDetailedErrorList(outputDir.getAbsolutePath(), inputFile + ",   " + e.getMessage(), bForceLogFiles);
 
                     //                   System.exit(1);
                 }
             }
-            CLIStandalone.writeLogForOmexCreation(outputDir, hasNonSpatialSet, hasSpatialSet);
+			writeLogForOmexCreation(outputDir.getAbsolutePath(), hasNonSpatialSet, hasSpatialSet, bForceLogFiles);
         } else {
             try {
                 assert input != null;
                 if (input.toString().endsWith(".vcml")) {
-                    boolean isCreated = vcmlToOmexConversion(cliHandler.getInputFilePath(), null, cliDatabaseService, cliHandler, hasNonSpatialSet, hasSpatialSet);
-                    if (isCreated) System.out.println("Combine archive created for `" + input + "`");
-                    else System.err.println("Failed converting VCML to OMEX archive for `" + input + "`");
-                } else System.err.println("No input files found in the directory `" + input + "`");
+                    boolean isCreated = vcmlToOmexConversion(input.getAbsolutePath(), null,
+							outputDir.getAbsolutePath(), cliDatabaseService, hasNonSpatialSet, hasSpatialSet,
+							modelFormat, bHasDataOnly, bNonSpatialOnly, bMakeLogsOnly, bForceLogFiles);
+                    if (isCreated) {
+						System.out.println("Combine archive created for `" + input + "`");
+					} else {
+						System.err.println("Failed converting VCML to OMEX archive for `" + input + "`");
+					}
+                } else {
+					System.err.println("No input files found in the directory `" + input + "`");
+				}
             } catch (Exception e) {
 //                e.printStackTrace();
                 System.out.println("\n\n\n=====>>>>EXPORT FAILED: " +input+"\n\n\n");   
@@ -175,9 +183,14 @@ public class VcmlOmexConverter {
         }
     }
 
-    private static boolean vcmlToOmexConversion(String inputFilePath, String outputBaseDir,
-												CLIDatabaseService cliDatabaseService, CLIHandler cliHandler,
-												Set<String> hasNonSpatialSet, Set<String> hasSpatialSet
+    private static boolean vcmlToOmexConversion(String inputFilePath, String outputBaseDir, String outputDir,
+												CLIDatabaseService cliDatabaseService,
+												Set<String> hasNonSpatialSet, Set<String> hasSpatialSet,
+												ModelFormat modelFormat,
+												boolean bHasDataOnly,
+												boolean bNonSpatialOnly,
+												boolean bMakeLogsOnly,
+												boolean bForceLogFiles
 	) throws XmlParseException, IOException, DataAccessException, SQLException {
 
         // Get VCML file path from -i flag
@@ -186,8 +199,6 @@ public class VcmlOmexConverter {
         
         String inputVcmlFile = inputFilePath;
 
-        // Get directory file path from -o flag
-        String outputDir = cliHandler.getOutputDirPath();
 
         // get VCML name from VCML path
         //String vcmlName = inputVcmlFile.split(File.separator, 10)[inputVcmlFile.split(File.separator, 10).length - 1].split("\\.", 5)[0];
@@ -208,7 +219,7 @@ public class VcmlOmexConverter {
         
         int numSimulations = bioModel.getNumSimulations();
         if(outputBaseDir != null && numSimulations == 0) {
-        	CLIStandalone.writeSimErrorList(outputBaseDir, vcmlName + " has no simulations.");
+			writeSimErrorList(outputBaseDir, vcmlName + " has no simulations.", bForceLogFiles);
         	return false;
         }
         bioModel.refreshDependencies();
@@ -265,7 +276,7 @@ public class VcmlOmexConverter {
 			}
 			if(bFieldDataFound) {
 				simulationsToRemove.add(simulation);
-	        	CLIStandalone.writeSimErrorList(outputBaseDir, vcmlName + " excluded: FieldData not supported at this time.");
+				writeSimErrorList(outputBaseDir, vcmlName + " excluded: FieldData not supported at this time.", bForceLogFiles);
 				SolverDescription solverDescription = simulation.getSolverTaskDescription().getSolverDescription();
 				String solverName = solverDescription.getShortDisplayLabel();
 				//CLIStandalone.writeSimErrorList(outputBaseDir, "   " + solverName);
@@ -295,11 +306,11 @@ public class VcmlOmexConverter {
         // NOTE: SEDML exporter exports both SEDML as well as required SBML
         List<Simulation> simsToExport = new ArrayList<Simulation>();
         LinkedHashSet<String> solverNames = new LinkedHashSet<>();
-		boolean bForceSBML = cliHandler.shouldForceSbml();
-		boolean bForceVCML = cliHandler.shouldForceVcml();
-		boolean bHasDataOnly = cliHandler.isHasDataOnly();
-		boolean bMakeLogsOnly = cliHandler.isMakeLogsOnly();
-		boolean bNonSpatialOnly = cliHandler.isNonSpacialOnly();
+//		boolean bForceSBML = cliHandler.shouldForceSbml();
+//		boolean bForceVCML = cliHandler.shouldForceVcml();
+//		boolean bHasDataOnly = cliHandler.isHasDataOnly();
+//		boolean bMakeLogsOnly = cliHandler.isMakeLogsOnly();
+//		boolean bNonSpatialOnly = cliHandler.isNonSpacialOnly();
 
         if (bHasDataOnly) {
         	// make list of simulations to export with only sims that have data on the server
@@ -342,19 +353,19 @@ public class VcmlOmexConverter {
         if(outputBaseDir != null && bHasDataOnly == true && simsToExport.size() == 0) {
         	String msg = vcmlName + " has no simulations with any results.";
         	System.out.println(msg);
-        	CLIStandalone.writeSimErrorList(outputBaseDir, msg);
+			writeSimErrorList(outputBaseDir, msg, bForceLogFiles);
         	return false;
         }
         if(outputBaseDir != null && bNonSpatialOnly == true && simsToExport.size() == 0) {
         	String msg = vcmlName + " has no non-spatial simulations to export.";
         	System.out.println(msg);
-        	CLIStandalone.writeSimErrorList(outputBaseDir, msg);
+			writeSimErrorList(outputBaseDir, msg, bForceLogFiles);
         	return false;
         }
         if(outputBaseDir != null && simsToExport.size() == 0) {
         	String msg = vcmlName + " has no simulations to export.";
         	System.out.println(msg);
-        	CLIStandalone.writeSimErrorList(outputBaseDir, msg);
+			writeSimErrorList(outputBaseDir, msg, bForceLogFiles);
         	return false;
         }
         
@@ -389,13 +400,15 @@ public class VcmlOmexConverter {
         XmlUtil.writeXMLStringToFile(rdfString, String.valueOf(Paths.get(outputDir, "metadata.rdf")), true);
         
         SEDMLExporter sedmlExporter = new SEDMLExporter(bioModel, sedmlLevel, sedmlVersion, simsToExport);
-        SEDMLDocument sedmlDocument = sedmlExporter.getSEDMLFile0(outputDir, vcmlName, bForceVCML, bForceSBML, bHasDataOnly, true);
+        SEDMLDocument sedmlDocument = sedmlExporter.getSEDMLFile0(outputDir, vcmlName,
+				modelFormat.equals(ModelFormat.VCML), modelFormat.equals(ModelFormat.SBML),
+				bHasDataOnly, true);
         SedML sedmlModel = sedmlDocument.getSedMLModel();
         if(sedmlModel.getModels().size() == 0) {
             File dir = new File(outputDir);
             String[] files = dir.list();
             removeOtherFiles(outputDir, files);
-        	CLIStandalone.writeSimErrorList(outputBaseDir, vcmlName + ": the sedm is empty.");
+			writeSimErrorList(outputBaseDir, vcmlName + ": the sedm is empty.", bForceLogFiles);
 //        	String allSolverNames = "";
 //        	for(String solverName : solverNames) {
 //        		allSolverNames += (solverName + " ");
@@ -773,4 +786,48 @@ public class VcmlOmexConverter {
 		System.out.println("finished");
 */
     }
+
+	// when creating the omex files from vcml, we write here the list of models that have spatial, non-spatial or both applications
+	public static void writeLogForOmexCreation(String outputBaseDir, Set<String> hasNonSpatialSet, Set<String> hasSpatialSet, boolean bForceLogFiles) throws IOException {
+		if (CLIUtils.isBatchExecution(outputBaseDir, bForceLogFiles)) {
+			String s = "";
+			s += "Only Non-spatial applications\n";
+			for (String name : hasNonSpatialSet) {
+				if (!hasSpatialSet.contains(name)) {
+					s += (name + "\n");
+				}
+			}
+			s += "\nOnly Spatial applications\n";
+			for (String name : hasSpatialSet) {
+				if (!hasNonSpatialSet.contains(name)) {
+					s += (name + "\n");
+				}
+			}
+			s += "\nBoth Spatial and Non-Spatial applications\n";
+			Set<String> hasBothSet = new LinkedHashSet<>(hasSpatialSet);
+			hasBothSet.addAll(hasNonSpatialSet);
+			for (String name : hasBothSet) {
+				s += (name + "\n");
+			}
+			String dest = outputBaseDir + File.separator + "omexCreationLog.txt";
+			Files.write(Paths.get(dest), (s + "\n").getBytes(),
+					StandardOpenOption.CREATE, StandardOpenOption.APPEND);
+		}
+	}
+	public static void writeMultiModelPublications(String outputBaseDir, String s, boolean bForceLogFiles) throws IOException {
+		if (CLIUtils.isBatchExecution(outputBaseDir, bForceLogFiles)) {
+			String dest = outputBaseDir + File.separator + "multiModelPublications.txt";
+			Files.write(Paths.get(dest), (s + "\n").getBytes(),
+					StandardOpenOption.CREATE, StandardOpenOption.APPEND);
+		}
+	}
+	// biomodels with no simulations and biomodels with no sim results (fired when building the omex files)
+	public static void writeSimErrorList(String outputBaseDir, String s, boolean bForceLogFiles) throws IOException {
+		if (CLIUtils.isBatchExecution(outputBaseDir, bForceLogFiles)) {
+			String dest = outputBaseDir + File.separator + "simsErrorLog.txt";
+			Files.write(Paths.get(dest), (s + "\n").getBytes(),
+					StandardOpenOption.CREATE, StandardOpenOption.APPEND);
+		}
+	}
+
 }
