+++ vcell-sbmlsim/pom.xml
@@ -123,7 +123,7 @@
 		<dependency>
 			<groupId>info.picocli</groupId>
 			<artifactId>picocli</artifactId>
-			<version>3.0.0-alpha-4</version>
+			<version>4.6.3</version>
 		</dependency>
 
 	</dependencies>
